% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapBayesVSpline.R
\name{adapBayesVSpline}
\alias{adapBayesVSpline}
\title{Bayes estimate for adaptive V-splines}
\usage{
adapBayesVSpline(X, Y, V, coff, w, gamma, est)
}
\description{
Fitting the Bayes estimate of adaptive V-splines
}
