% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapCofBayesVSpline.R
\name{adapCofBayesVSpline}
\alias{adapCofBayesVSpline}
\title{Coefficients for adaptive V-splines}
\usage{
adapCofBayesVSpline(X, Y, V, lambs, gam)
}
\description{
To reconstruct the coefficients for adaptive V-splines.
}
