% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitAdapBayesVSpline.R
\name{fitAdapBayesVSpline}
\alias{fitAdapBayesVSpline}
\title{Bayes estimate of adaptive V-Splines}
\usage{
fitAdapBayesVSpline(
  dat,
  W = NULL,
  U = NULL,
  pa = NULL,
  a = 3,
  b = 2,
  xout = NULL
)
}
\description{
Fitting adaptive V-splines with piecewise \eqn{\lambda}s.
}
