% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitBayesVSpline.R
\name{fitBayesVSpline}
\alias{fitBayesVSpline}
\title{Bayes estimate of V-Spline model}
\usage{
fitBayesVSpline(dat, W = NULL, U = NULL, pa = NULL, xout = NULL)
}
\description{
Bayes estimate of V-Spline model
}
