% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitVSP.R
\name{fitVSP}
\alias{fitVSP}
\title{Fit V-Spline model}
\usage{
fitVSP(dat, pa = NULL, xout = NULL, a = 3, b = 2)
}
\description{
Fit V-Spline model
}
\examples{
n <-100
s <- seq(1,2*pi,length=n)
y <- sin(s) + rnorm(n)
v <- cos(s) + rnorm(n)
simuData <- data.frame(t=s,y=y,v=v)
fitted   <- fitVSP(simuData)
plot(s,y)
points(fitted$vsp$t,fitted$vsp$y,col="red",type="l")
}
\references{
Z. Cao, D. Bryant, C. Fox, T. Molten and M. Parry. "V-Spline: an Adaptive Smoothing Spline
for Trajectory Reconstruction" arXiv preprint arXiv:1803.07184 (2018).
}
\seealso{
\code{\link{getTheta}}, \code{\link{loocvVSP}}
}
