% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iniVSPmatrices.R
\name{iniVSPmat}
\alias{iniVSPmat}
\title{Initialise V-Spline matrices \eqn{B}, \eqn{C} and \eqn{\Omega}}
\usage{
iniVSPmat(dat, a = 3, b = 2)
}
\description{
This function is utilised for initializing the V-Spline matreces \eqn{B},
\eqn{C} and \eqn{\Omega}, with whitch it potentially saves computating time.

The default values of \eqn{a} and \eqn{b} are 3 and 2 respectively.
}
\keyword{V-Spline}
\keyword{matrices}
