% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapKernel.R
\name{kernelR1adap}
\alias{kernelR1adap}
\alias{dotR1adap}
\alias{dR1adap}
\alias{ddotR1adap}
\alias{ddotdotR1adap}
\title{Kernel functions of Bayes estimate of V-Splines}
\usage{
kernelR1adap(x, y, w, alt)

dotR1adap(x, y, w, alt)

dR1adap(x, y, w, alt)

ddotR1adap(x, y, w, alt)

ddotdotR1adap(x, y, w, alt)
}
\description{
adaptive Kernel functions \eqn{R_w^1(s,t), \dot{R}_w^1(s,t), 
R_w^{'1}(s,t), \dot{R}_w^{'1}(s,t)}.
}
\references{
Z. Cao, D. Bryant, C. Fox, T. Molten and M. Parry. "V-Spline: an Adaptive Smoothing Spline
for Trajectory Reconstruction" arXiv preprint arXiv:1803.07184 (2018).

Z. Cao, D. Bryant, and M. Parry. "V-spline and its Bayes estimate" arXiv (2018).
}
\keyword{Kernel}
\keyword{adaptive}
\keyword{functions}
