% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loocvVSP.R
\name{loocvVSP}
\alias{loocvVSP}
\title{Leave-one-out corss validation for V-Spline}
\usage{
loocvVSP(dat, pa, matrix_list)
}
\arguments{
\item{dat}{the dataframe}

\item{pa}{the vector of the parameters \eqn{\lambda} and \eqn{\eta}}

\item{matrix_list}{the matrix list comprises \eqn{B}, \eqn{C} and \eqn{\Omega}, which are obtained from \code{iniVSPmatrics()}}
}
\description{
The leave-one-out corss validation is used for finding the optimal parameters for V-Spline.
}
\keyword{corss}
\keyword{leave-one-out}
\keyword{validation}
