% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcstVSP.R
\name{rcstVSP}
\alias{rcstVSP}
\title{Reconstruction points}
\usage{
rcstVSP(data_frame, theta, xout = NULL)
}
\description{
To reconstruct the V-Spline with \eqn{\theta}, which is gained from \code{getTheta()}.
}
\references{
Z. Cao, D. Bryant, C. Fox, T. Molten and M. Parry. "V-Spline: an Adaptive Smoothing Spline
for Trajectory Reconstruction" arXiv preprint arXiv:1803.07184 (2018).
}
