% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimuData.R
\name{SimuData}
\alias{SimuData}
\title{Generate simulation data}
\usage{
SimuData(v, rsnr)
}
\arguments{
\item{v}{noise-free velocity data}

\item{rsnr}{signal to noise ratio
SimuData()}
}
\description{
This function takes *noise-free* velocity data
computes the implied position data and then adds noise.
It requires function forward1.
}
\keyword{V-Spline}
\keyword{basis}
\keyword{functions#'}
