% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basisFunctions.R
\name{bN1}
\alias{bN1}
\title{V-Spline basis functions}
\usage{
bN1(k, t, est)
}
\arguments{
\item{k}{to determine the location of the estimate.}

\item{t}{Time sequence.}

\item{est}{The target estimate.}
}
\description{
V-Spline basis functions allow to you build a V-Spline.
The basis functions are denoted by \eqn{N_1(t), N_2(t), \ldots, N_{2n}(t)}.
}
\references{
Z. Cao, D. Bryant, C. Fox, T. Molten and M. Parry. "V-Spline: an Adaptive Smoothing Spline
for Trajectory Reconstruction" arXiv preprint arXiv:1803.07184 (2018).
}
\keyword{V-Spline}
\keyword{basis}
\keyword{functions}
