% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitAdapBayesVSpline.R
\name{fitAdapBayesVSpline}
\alias{fitAdapBayesVSpline}
\title{Bayes estimate of adaptive V-Splines}
\usage{
fitAdapBayesVSpline(dat, W = NULL, U = NULL, pa = c(1, 1),
  xout = NULL)
}
\description{
Bayes estimate of adaptive V-Splines
}
