% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimuData.R
\name{SimuData}
\alias{SimuData}
\title{Generate simulation data}
\usage{
SimuData(v, rsnr)
}
\arguments{
\item{v}{noise-free velocity data}

\item{rsnr}{signal to noise ratio}
}
\description{
This function takes *noise-free* velocity data
computes the implied position data and then adds noise.
It requires function forward1.
}
\examples{
library(waveband)
library(VSPline)
set.seed(2016)
N <- 100
velocity <- test.data(type = "blocks", n = N, signal = 1, rsnr = 7, plotfn = TRUE)
set.seed(2016)
position <- SimuData(velocity$y,7)
simulatedata<- data.frame(t=velocity$x,y=position$xnoise,v=velocity$ynoise,boom=0)
}
\keyword{V-Spline}
\keyword{basis}
\keyword{functions#'}
